Enable GRPC transaction and query options (#7439)
## Release notes: product changes
Enable options provided by the TypeDB GRPC protocol.

**Enable transaction options:**
* `transaction_timeout`: If set, specifies a timeout for killing
transactions automatically, preventing memory leaks in unclosed
transactions,
* `schema_lock_acquire_timeout`: If set, specifies how long the driver
should wait if opening a transaction is blocked by an exclusive schema
write lock.

**Enable query options:**
* `include_instance_types`: If set, specifies if types should be
included in instance structs returned in ConceptRow answers,
* `prefetch_size`: If set, specifies the number of extra query responses
sent before the client side has to re-request more responses. Increasing
this may increase performance for queries with a huge number of answers,
as it can reduce the number of network round-trips at the cost of more
resources on the server side.

## Motivation
Allow default transaction and query behaviors to be overridden by GRPC
clients.

## Implementation
While the logic has already been propagated to the whole GRPC
transaction service, there were desired refinements of the protocol
structure's processing. Additionally, to ensure that the BDD tests
updates are in sync with the server, the relevant Bazel references are
updated.

+++ tests/behaviour/service/http/http_steps/query.rs
@@ -271,7 +271,7 @@ pub async fn set_query_option_include_instance_types(context: &mut Context, valu
 
 #[apply(generic_step)]
 #[step(expr = "set query option answer_count_limit to: {int}")]
-pub async fn set_query_option_answer_count_limit(context: &mut Context, value: usize) {
+pub async fn set_query_option_answer_count_limit(context: &mut Context, value: u64) {
     context.init_query_options_if_needed();
     context.query_options.as_mut().unwrap().answer_count_limit = Some(value);
 }
